prescription-module/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
python-multipart>=0.0.6
